refactor(harness): own extension and event facades
