+++ src/loushang/harness/extensions/__init__.py
@@ -4,4 +4,6 @@ from loushang.harness.extensions.provider_runtime import (
     ExtensionProviderRuntime,
     ProviderFactory,
 )
-__all__ = ["ExtensionProviderRuntime", "ProviderFactory"]
+from loushang.harness.extensions.runner import ExtensionRunner
+
+__all__ = ["ExtensionProviderRuntime", "ExtensionRunner", "ProviderFactory"]
